requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
bcrypt==4.0.1
python-jose
cryptography